fix: wire connection under azd ai agent (per Travis guidance)
Revert namespace change from ai to ai.agent. Connection commands now live
at azd ai agent connection until the azd core namespace change lands.

Per Travis: write commands under azd ai agent for now, keep code in
separate files/packages for easy lift-and-shift when Jeffrey's core
change is ready.

Changes:
- Revert extension.yaml namespace to ai.agent
- Revert main.go to original import
- Add connection command to existing cmd/root.go NewRootCommand
- Remove unused internal/root.go and exported SetupDebugLogging

The connection code stays self-contained in internal/connections/
with no imports from internal/cmd/ (agent code).
