+++ cli/azd/extensions/azure.ai.agents/extension.yaml
@@ -1,9 +1,9 @@
 # yaml-language-server: $schema=../extension.schema.json
 id: azure.ai.agents
-namespace: ai
-displayName: Foundry AI (Preview)
-description: Manage agents and connections in Microsoft Foundry. (Preview)
-usage: azd ai <command> [options]
+namespace: ai.agent
+displayName: Foundry agents (Preview)
+description: Ship agents with Microsoft Foundry from your terminal. (Preview)
+usage: azd ai agent <command> [options]
 # NOTE: Make sure version.txt is in sync with this version.
 version: 0.1.31-preview
 requiredAzdVersion: ">1.23.13"
